Add cache x86 proof

+++ timing/examples/x86_wcet_comparison/cache/cache.c
@@ -1,7 +1,8 @@
 #include <stdlib.h>
 #include <stdio.h>
+#include <stdint.h>
 
-int sum(int* arr, int size) {
+int sum(int* arr, uint64_t size) {
     int x = 0;
     for (int i = 0; i < size; i++) {
         x += arr[i];
